+++ showimage.py
@@ -21,7 +21,7 @@ def main():
     rectlist = []
     displaypar = 1
     step = 10
-    displaypar = int(tree.find("display-par").text)
+    displaypar = float(tree.find("display-par").text)
     step = tree.find("step").text
     for elem in tree.iter(tag='rect'):
         tempitem = RectItem()
